+++ CHANGELOG.md
@@ -2,11 +2,17 @@
 
 All notable changes to this project will be documented in this file, in reverse chronological order by release.
 
-## 1.0.1 - TBD
+## 1.1.0 - 2018-09-12
 
 ### Added
 
-- Nothing.
+- [#27](https://github.com/zendframework/zend-expressive-session/pull/27) adds a new interface, `Zend\Expressive\Session\SessionIdentifierAwareInterface`.
+  `SessionInterface` implementations should also implement this interface, and
+  persistence implementations should only create and consume session
+  implementations that implement it. The interface defines a single method,
+  `getId()`, representing the identifier of a discovered session. This allows
+  the identifier to be associated with its session data, ensuring that when
+  concurrent requests are made, persistence operates on the correct identifier.
 
 ### Changed